+++ src/plugins/e-acsl/tests/e-acsl-runtime/oracle/gen_typedef.c
@@ -6,8 +6,8 @@ int main(void)
   uint8 x;
   x = (unsigned char)0;
   /*@ assert x ≡ 0; */
-  __e_acsl_assert((int)x == 0,(char *)"Assertion",(char *)"main",
-                  (char *)"x == 0",9);
+  __e_acsl_assert(x == 0,(char *)"Assertion",(char *)"main",(char *)"x == 0",
+                  9);
   __retres = 0;
   return __retres;
 }
